Repairs to LunEx libraries
Changed LunEx currency from double to an int.
Moved classes out of default package.

+++ LunEx/src/ITC/SecurityExchangeTransmissionInterface.java
@@ -1,5 +1,5 @@
 package ITC;
 
 public interface SecurityExchangeTransmissionInterface {
-	double currentPrice(String symbol);
+	int currentPrice(String symbol);
 }
